move a lot of logic out of iceberg_multi_file_list.cpp
