+++ src/core/deletes/CMakeLists.txt
@@ -1,7 +1,5 @@
-add_library(
-  iceberg_core_deletes OBJECT
-  iceberg_deletion_vector.cpp iceberg_equality_delete.cpp
-  iceberg_positional_delete.cpp)
+add_library(iceberg_core_deletes OBJECT iceberg_deletion_vector.cpp
+                                        iceberg_positional_delete.cpp)
 set(ALL_OBJECT_FILES
     ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:iceberg_core_deletes>
     PARENT_SCOPE)
