+++ src/core/deletes/iceberg_deletion_vector.cpp
@@ -6,14 +6,6 @@
 #include "duckdb/common/string_util.hpp"
 #include "duckdb/common/vector_operations/unary_executor.hpp"
 
-#include "planning/iceberg_multi_file_list.hpp"
-#include "catalog/rest/catalog_entry/table/iceberg_table_information.hpp"
-#include "catalog/rest/api/catalog_utils.hpp"
-#include "core/metadata/puffin/iceberg_puffin_metadata.hpp"
-#include "iceberg_options.hpp"
-
-#include <variant>
-
 namespace duckdb {
 
 namespace {
@@ -138,96 +130,6 @@ shared_ptr<IcebergDeletionVectorData> IcebergDeletionVectorData::FromBlob(const 
 		    stored_checksum, checksum);
 	}
 	return result_p;
-}
-
-//! Verify a deletion-vector file is a structurally-valid Puffin container before reading it. The
-//! footer payload must be uncompressed (checked unconditionally): the Puffin spec permits an
-//! LZ4-compressed footer, but a deletion-vector footer must be plain so any reader walking the
-//! footer can parse the FileMetadata without a decompressor. In debug builds we additionally parse
-//! the whole container -- leading/footer/trailing magic, the footer JSON, and that the single
-//! blob's offset/length agree with the manifest content_offset/content_size_in_bytes used below.
-using PuffinDeletionVectorVerificationResult = std::variant<std::monostate, string>;
-
-static PuffinDeletionVectorVerificationResult VerifyPuffinDeletionVector(FileSystem &fs, FileHandle &handle,
-                                                                         int64_t content_offset, int64_t content_size) {
-	auto footer_result = IcebergPuffinReader::ReadFooter(fs, handle, handle.GetPath());
-	if (auto error = std::get_if<string>(&footer_result)) {
-		return *error;
-	}
-	auto &footer = std::get<IcebergPuffinFileFooter>(footer_result);
-
-	bool contains_blob = false;
-	for (auto &blob : footer.file_metadata.blobs) {
-		if (blob.type != "deletion-vector-v1") {
-			return "Deletion vector Puffin blob type mismatch: expected deletion-vector-v1";
-		}
-		if (blob.offset != content_offset || blob.length != content_size) {
-			continue;
-		}
-		contains_blob = true;
-	}
-	if (!contains_blob) {
-		return StringUtil::Format("Deletion vector blob with offset (%d) and length (%d) not found in Puffin file",
-		                          content_offset, content_size);
-	}
-	return std::monostate {};
-}
-
-void IcebergMultiFileList::ScanPuffinFile(const BoundIcebergManifestEntry &bound_entry) const {
-	auto &entry = bound_entry.entry;
-	auto &data_file = entry.data_file;
-	auto &table_metadata = GetMetadata();
-	auto iceberg_version = table_metadata.iceberg_version;
-	if (iceberg_version < 3) {
-		throw InvalidConfigurationException("DeletionVector not supported in Iceberg V%d", iceberg_version);
-	}
-	auto file_path = data_file.file_path;
-	if (!data_file.referenced_data_file) {
-		throw InvalidConfigurationException("Puffin delete file is missing 'referenced_data_file'");
-	}
-
-	FileOpenFlags flags = FileFlags::FILE_FLAGS_READ;
-	flags.SetCachingMode(CachingMode::CACHE_REMOTE_ONLY);
-	auto file_handle = fs.OpenFile(file_path, flags);
-
-	if (!data_file.content_offset) {
-		throw InvalidConfigurationException("Puffin delete file is missing 'content_offset");
-	}
-	if (!data_file.content_size_in_bytes) {
-		throw InvalidConfigurationException("Puffin delete file is missing 'content_size_in_bytes");
-	}
-
-	auto offset = *data_file.content_offset;
-	auto length = *data_file.content_size_in_bytes;
-
-	auto local_buffer = Allocator::DefaultAllocator().Allocate(length);
-	fs.Read(*file_handle, local_buffer.get(), length, offset);
-
-	Value skip_verification;
-	if (!context.TryGetCurrentSetting(SKIP_PUFFIN_VERIFICATION_CONFIG_VARIABLE, skip_verification) ||
-	    !skip_verification.GetValue<bool>()) {
-		auto verification_result = VerifyPuffinDeletionVector(fs, *file_handle, offset, length);
-		if (auto error = std::get_if<string>(&verification_result)) {
-			throw InvalidConfigurationException("%s. Older versions of DuckDB wrote deletion vector files as bare "
-			                                    "blobs. To read those files, run \"SET "
-			                                    "%s = true\"",
-			                                    *error, SKIP_PUFFIN_VERIFICATION_CONFIG_VARIABLE);
-		}
-	}
-
-	auto &positional_delete_data = GetPositionalDeleteData();
-	auto it = positional_delete_data.find(*data_file.referenced_data_file);
-	if (it != positional_delete_data.end()) {
-		//! Another delete already exists for this table
-		auto &existing_delete = *it->second;
-		if (existing_delete.type == IcebergDeleteType::DELETION_VECTOR) {
-			throw InvalidConfigurationException(
-			    "Table is corrupt, two or more deletion vectors exist for the same referenced_data_file");
-		}
-	}
-	//! NOTE: assign, don't emplace, deletion vectors take priority over any remaining positional delete files
-	positional_delete_data[*data_file.referenced_data_file] =
-	    IcebergDeletionVectorData::FromBlob(bound_entry, local_buffer.get(), length);
 }
 
 idx_t IcebergDeletionVector::Filter(row_t start_row_index, idx_t count, SelectionVector &result_sel) {
