+++ src/include/planning/deletes/iceberg_delete_file_scanner.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "planning/deletes/iceberg_delete_planner.hpp"
+
+namespace duckdb {
+
+struct IcebergDeleteFileScanner {
+	static void ScanFiles(const IcebergDeletePlanningContext &context);
+};
+
+} // namespace duckdb
