+++ src/planning/scan_order/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_library(iceberg_planning_scan_order OBJECT iceberg_scan_order.cpp)
+set(ALL_OBJECT_FILES
+    ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:iceberg_planning_scan_order>
+    PARENT_SCOPE)
